requirements.txt
fastapi
uvicorn
requests
python-dotenv
chromadb
sentence-transformers
groq
streamlit
apscheduler
python-dotenv
feedparser
google-adk